Refactor: Simplify event classes
